+++ jdaviz/core/loaders/resolvers/__init__.py
@@ -4,3 +4,4 @@ from .file import *  # noqa
 from .file_drop import *  # noqa
 from .url import *  # noqa
 from .object import *  # noqa
+from .virtual_observatory import *  # noqa