+++ jdaviz/core/user_api.py
@@ -214,7 +214,12 @@ class LoaderUserApi(UserApiWrapper):
     """
     def __init__(self, loader, expose=[], readonly=[], excl_from_dict=[], deprecated=[]):
         expose = list(set(list(expose) + ['format', 'target', 'importer', 'load', 'show',
-                                          'open_in_tray', 'close_in_tray']))
+                                          'open_in_tray', 'close_in_tray',
+                                          'treat_table_as_query',
+                                          'observation_table', 'file_table',
+                                          'file_cache',
+                                          'file_timeout',
+                                          'file_local_path']))
         super().__init__(loader, expose, readonly, excl_from_dict, deprecated)
 
     def __repr__(self):