+++ CHANGES.rst
@@ -14,6 +14,9 @@ New Features
 - Viewers now support being popped out into separate window or sending to any sidecar location via `.show()`
   API. [#3821]
 
+- Loaders now all support "observation" and "file" tables, allowing choosing which file to download
+  and import from an astroquery or MAST search. [#3769]
+
 Cubeviz
 ^^^^^^^
 
@@ -21,6 +24,8 @@ Imviz
 ^^^^^
 
 - Loading footprints from files/API is now done through the loaders infrastructure. [#3790]
+
+- Virtual Observatory (VO) plugin functionality is now available through the loaders infrastructure. [#3769]
 
 Mosviz
 ^^^^^^